modify decommission
